+++ CHANGELOG.md
@@ -2,10 +2,20 @@
 
 All notable changes to the "under30" extension will be documented in this file.
 
-### 1.0.3
+### [1.1.0] - 2024-07-29
+
+## Added
+- New command `under30.setFunctionLengthLimit` to allow users to dynamically change the function length limit.
+- Updated extension to check function lengths based on the user-defined limit.
+
+### [1.0.3] - 2024-06-05
 
 Updated documentation and corrections for under-30 VSCode Extension.
 
-### 1.0.0
+### [1.0.2] - 2024-06-06
+
+Updated documentation and corrections for under-30 VSCode Extension.
+
+### [1.0.0] - 2024-06-05
 
 Initial release of under-30 VSCode Extension.